Backport support for multiple audio capture in decklink
Original commit:

decklink: Add support for capturing multiple audio streams

In most cases an SDI link will be composed of one or more pairs of
stereo audio.  Add support for capturing those pairs, each as a
separate audio stream.  This allows the user to choose which pair
to listen to in the standard audio stream dropdown box.

Signed-off-by: Devin Heitmueller <dheitmueller@ltnglobal.com>