Backport fix for lost 708 packets
Original commit:

Fix case where CDPs are being lost in H.264 streams

There are cases where the packetizer will be able to extract CDPs
from SCTE-128, but ultimately not return a block.  We need to
check for CC one last time at the end of the loop or else we these
CDPs will get discarded at the next flush.

Signed-off-by: Devin Heitmueller <dheitmueller@ltnglobal.com>