Backport code to use libklvanc for decklink capture
Original commit:

Bolt libklvanc into decklink capture

The CC extraction code built into VLC has a number of deficiencies,
such as not not handling SD VANC data, only working if the CC
packet is at the beginning of the line, and not handling a variety
of edge cases that can occur parsing the CDP.

Hook in libklvanc into VLC, so we can use it not just for CC but
for other VANC data types as well (e.g. AFD, SCTE-104, etc).

Signed-off-by: Devin Heitmueller <dheitmueller@ltnglobal.com>